Adds new check to see if an element is covered by another
(This message is written by @smblott-github.)

This is @poacher2k's #2251 rebased, tweaked and squashed into a single
commit.  It provides a new method of detecting when an overlay hides a
clickable element.

There is some possibility that this may have to be reverted (for
example, if we find we're missing too many important hints).  For that
reason, I squashed this into a single commit.

Closes #2251.

Adds another check to find out if elem is visible

Also adds 0.1 pixels inwards to corner coordinates to handle some edge
cases where another element would be found instead

Removes redundant contains-check

Reverts change that made tests fail

Tweak #2251 (better overlay detection).