Do not hardwire <c-[>; users can now unmap it.
Previously, <c-[> was hardwired into KeyboardUtils.isEscape().

This now simply makes <c-[> a regular default binding.

The advantage of doing this is that users can now unmap it.

There is a catch, however.  Previously, <c-[> was unaffected by
unmapAll.  Because it's now a regular binding, it would be removed by
unmapAll.  Therefore, to avoid changing the established behaviour, we
retain this binding even with unmapAll.