Fix tests and reorganise Chrome stubs.
The tests were passing but throwing errors because
`KeyboardUtils.isEscape()` now depends on `Settings`, and `Settings`
depends on various `chrome.*` APIs.

Fixing this required a bit of a reorganisation.  We need all of the
above in the Vomnibar and the HUD.  They're loaded from the
corresponding HTML files.  But those files are also loaded by the tests.
So we need a way to have those files load chrome stubs (for the tests)
if necessary, but not load them otherwise.

The solution (here) is to move the chrome stubs to
lib/chrome_stubs.coffee and include them from the two HTML files.  This
has no effect in a live installation.