Exit grab-back-focus on link hints.
Link hints works in all frames and grab-back-focus is active in every
frame.

Because the user is only interacting with one frame, only that frame
recognises that the user is interacting with the page.

As a result, if the user selects an input in another frame,
grab-back-focus immediately blurs the input.

Here, we exit all grab-back-focus modes when link hints launches.