Add mark option for go-to-mark.
Examples:

    map a Marks.activateGotoMode mark=A
    map q Marks.activateGotoMode mark=~ global

These go directly to the indicated mark.

For capital letters, we can differentiate global and local marks.  For
other characters, the use has to add the option `global` to instruct
Vimium to use a global mark.

This is likely to be most useful with global marks.

Helps with #2298 (but doesn't really fix it).