Use event.code to detect/suppress keyup events.
This avoids the possibility of leaking keyup events if the keys a
released in a different order from that in which they were pressed.

Also, replace suppressKeyupAfterEscape with this same mechanism.

This fixes a bug (in master/1.59) whereby we leak the keyup event for
`i` when entering insert mode.

TODO:

- `/`, `<Escape>` leaks a keyup event
- `i` leaks a keyup event