Fix search-engine completion (for firefox).
The problem is that Firefox throws a `DataCloneError` caused by
something in `response`.  That was being masked by the `try`.

The solution here is to just encode the entire response as JSON (and
decode it at the other end).

A better solution would be to figure out what (in `response`) is causing
the problem.

Fixes 2576.